requirements.txt
pandas
great_expectations
rich
pytest
pyyaml
openai
fastapi
uvicorn
